split out the get program data file
