+++ apps/valence-app/src/app/programs/server/actions/index.ts
@@ -3,12 +3,12 @@ export {
   type GetAllProgramsReturnValue,
 } from "./get-all-programs-from-registry";
 export {
-  type GetProgramDataReturnValue,
-  type FetchProcessorQueuesReturnType,
-  getProgramData,
-} from "./get-program-data";
-export {
   fetchLibrarySchema,
   type FetchLibrarySchemaReturnValue,
   jsonSchemaToTypescript,
 } from "./fetch-library-schema";
+export {
+  type GetProgramDataReturnValue,
+  type FetchProcessorQueuesReturnType,
+  getProgramData,
+} from "./get-program-data/main";
